Fix onComplete firing before Tween is done
The issue is that the tween doesn't run its last update until AFTER the
onComplete() handler is called. So not only does it mean that the
tweened value will be wrong in the onComplete, but it also means that if
you try to set that value to something in the handler, the tween will
overwrite whatever value you set.