Added config class for tween collections, fixes independent time scale issues with them.
Classes inheriting from AbstractGoTweenCollection were unable to make use of timeScaleIndependantUpdateType because this could only be set after the constructor was called (which adds the collection to the Go class update loop as a normal Update).

Using a config class allows this, and other parameters to be set upon construction in the same manner as the GoTween class thus making it more consistent, and fixing the timescale bug. Two birds, one stone.