A fix to stop multiple GoKit instances building up in a scene.
The issue is that as the application is quitting, the GoKit instance is destroyed and then GoKit.instance is referred to again, causing a new instance is created. This new instance is serialised with the scene and can't be found back with FindObjectOfType(). The side effect of multiple instances is that tweens run faster because they tick more then once a frame.

This fix simply marks when the application is quitting and doesn't allow another instance to be created.